pyproject.toml
[project]
name = "stats"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.2.2",
    "openai>=1.59.8",
    "python-dotenv>=1.0.1",
    "scipy>=1.15.1",
    "statsmodels>=0.14.4",
]
